Moved files to providers package.  Externalized connection logic from init function
