+++ airflow/providers/apache/kafka/sensors/kafka_sensor.py
@@ -10,6 +10,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import logging
+
 from cached_property import cached_property
 from airflow.utils.decorators import apply_defaults
 from airflow.sensors.base_sensor_operator import BaseSensorOperator
@@ -30,14 +32,13 @@ class KafkaSensor(BaseSensorOperator):
     @apply_defaults
     def __init__(self, topic, host=DEFAULT_HOST, port=DEFAULT_PORT, *args, **kwargs):
         """
-            Initialize the sensor, the connection establish
-            is put off to it's first time usage.
-
+        Initialize the sensor, the connection establish
+        is put off to it's first time usage.
+        :param conn_id:
+            the kafka broker connection whom this sensor
+            subscripts against.
         :param topic:
-        :param host:
-        :param port:
-        :param args:
-        :param kwargs:
+            the subscribed topic
         """
         self.topic = topic
         self.host = host
@@ -46,24 +47,18 @@ class KafkaSensor(BaseSensorOperator):
 
     @cached_property
     def hook(self):
-        """
-        Returns a Kafka Consumer Hook
-        """
         return KafkaConsumerHook(self.topic, self.host, self.port)
 
     def poke(self, context):
-        """
-            Checks to see if messages exist on this topic/partition.
-
-        :param context:
-        :return:
-        """
-        self.log.info('Poking topic: %s, using hook: %s', str(self.topic), str(self.hook))
+        logging.info(
+            'Poking topic: %s, using hook: %s',
+            self.topic, self.hook)
 
         messages = self.hook.get_messages()
 
-        if messages:
-            self.log.info('Got messages during poking: %s', str(messages))
+        if messages is not {}:
+            logging.info(
+                'Got messages during poking: %s', messages)
             return messages
         else:
             return False
